refactoring added info about custom
