+++ main.py
@@ -6,7 +6,7 @@ from os import getenv
 import asyncio
 from aiogram import Bot, Dispatcher
 
-from handlers import menu, different_types, order_status, info, business
+from handlers import menu, different_types, order_status, info
 
 
 # Запуск бота
@@ -14,7 +14,7 @@ async def main():
     token = getenv("BOT_TOKEN")
     bot = Bot(token=token)
     dp = Dispatcher()
-    dp.include_routers(menu.router, different_types.router, order_status.router, info.router, business.router)
+    dp.include_routers(menu.router, different_types.router, order_status.router, info.router)
 
     # Запускаем бота и пропускаем все накопленные входящие
     # Да, этот метод можно вызвать даже если у вас поллинг
